+++ entry.js
@@ -1,0 +1,10 @@
+// Import CSS
+// import '../css/master.scss';
+
+// Import React and JS
+import HelloBox from './HelloBox';
+import React from 'react';
+import ReactDom from 'react-dom';
+
+// Render!
+// ReactDom.render(<HelloBox />, document.getElementsByTagName('body')[0]);