+++ changelog.md
@@ -1,3 +1,9 @@
+#### 16.07.14
+
+- Finish the [version 1](./trello) for [Trello](http://trello.com)
+- Re-implement the **My Board** page & **Board Detail** page
+- Using ReactJS
+
 #### 16.07.12
 
 - init the project and start to write the first stunning website